requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
audioread==3.0.1
av==15.0.0
beautifulsoup4==4.13.4
boto3==1.39.10
botocore==1.39.10
Brotli==1.1.0
cached_path==1.7.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
deep-translator==1.11.4
dill==0.3.8
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
encodec==0.1.1
f5-tts @ git+https://github.com/VYNCX/F5-TTS-THAI.git@708be783350d22ecb997d3c8165106c708dc13e6
fastapi==0.116.1
faster-whisper==1.1.1
ffmpy==0.6.1
filelock==3.18.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.38.0
gradio_client==1.11.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.22.1
openai-whisper==20250625
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
pypinyin==0.55.0
pythainlp==5.1.2
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rsa==4.9.1
ruff==0.12.4
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
semantic-version==2.10.0
sentry-sdk==2.33.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
ssg==0.0.8
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.47.2
syllapy==0.7.2
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchdiffeq==0.2.5
tqdm==4.67.1
transformers==4.53.3
transformers-stream-generator==0.0.5
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vocos==0.1.0
wandb==0.21.0
websockets==15.0.1
x-transformers==2.4.12
xxhash==3.5.0
yarl==1.20.1
